requirements.txt
Django==1.10.6
Pillow==4.0.0
Unidecode==0.04.20
bbcode==1.0.24
blinker==1.3
boto==2.46.1
django-annoying==0.10.3
django-environ==0.4.1
django-registration-redux==1.4
django-storages==1.5.2
gunicorn==19.7.1
jsonpickle==0.9.2
olefile==0.44
psycopg2==2.7.1
pybbm==0.18.4
raygun4py==3.1.4
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
